+++ src/test/php/net/stubbles/webapp/RoutingTestCase.php
@@ -55,6 +55,16 @@ class RoutingTestCase extends \PHPUnit_Framework_TestCase
     /**
      * @test
      */
+    public function annotationPresentOnSetAuthHandlerMethod()
+    {
+        $method = lang\reflect($this->routing, 'setAuthHandler');
+        $this->assertTrue($method->hasAnnotation('Inject'));
+        $this->assertTrue($method->getAnnotation('Inject')->isOptional());
+    }
+
+    /**
+     * @test
+     */
     public function returnsMissingRouteOnRouteSelectionWhenNoRouteAdded()
     {
         $this->assertInstanceOf('net\stubbles\webapp\MissingRoute',
@@ -97,6 +107,31 @@ class RoutingTestCase extends \PHPUnit_Framework_TestCase
         $this->routing->onGet('/foo', function() {});
         $this->assertInstanceOf('net\stubbles\webapp\MethodNotAllowedRoute',
                                 $this->routing->findRoute($this->calledUri)
+        );
+    }
+
+    /**
+     * @test
+     * @since  3.0.0
+     */
+    public function returnsInternalServerErrorRouteWhenMatchingRouteRequiresAuthButNoAuthHandlerSet()
+    {
+        $this->routing->onGet('/hello', function() {})->withRoleOnly('admin');
+        $this->assertInstanceOf('net\stubbles\webapp\InternalServerErrorRoute',
+                                $this->routing->findRoute($this->calledUri)
+        );
+    }
+
+    /**
+     * @test
+     * @since  3.0.0
+     */
+    public function returnsAuthorizungRouteWhenMatchingRouteRequiresAuthButNoAuthHandlerSet()
+    {
+        $this->routing->onGet('/hello', function() {})->withRoleOnly('admin');
+        $this->assertInstanceOf('net\stubbles\webapp\auth\AuthorizingRoute',
+                                $this->routing->setAuthHandler($this->getMock('net\stubbles\webapp\auth\AuthHandler'))
+                                              ->findRoute($this->calledUri)
         );
     }