+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+3.0.0, (2013-11-01)
+-------------------
+
+### BC breaks
+
+   * interface `net\stubbles\webapp\AuthHandler` was replaced by `net\stubbles\webapp\auth\AuthHandler`
+
+### Other changes
+
+   * implemented #32: possibility to propagate an error in the auth system to the response
+
+
 2.2.0, (2013-10-13)
 -------------------